Rename viewer files for clarity: box_view2d.py and box_view3d.py
Files renamed:
  - map_box_view.py → box_view2d.py (2D solar map + box selector)
  - magfield_viewer.py → box_view3d.py (3D magnetic field visualization)

Changes:
  - Updated all import statements in __init__.py, fov_selector_gui.py,
    box_view2d.py (formerly map_box_view.py), and view_h5.py
  - Updated state filenames to box_view3d_state.pkl for consistency
  - Maintained class names (MapBoxDisplayWidget, MagFieldViewer) for
    semantic correctness
  - All tests passing

Benefits:
  - Symmetric naming clearly indicates paired 2D/3D viewers
  - Shorter, more intuitive file names
  - Improved code discoverability
