+++ pyampp/gxbox/view_h5.py
@@ -356,7 +356,7 @@ def prepare_model_for_viewer(model_path: str | Path) -> tuple[SimpleBox, Time, s
 
 
 def main() -> int:
-    from pyampp.gxbox.magfield_viewer import MagFieldViewer
+    from pyampp.gxbox.box_view3d import MagFieldViewer
 
     parser = argparse.ArgumentParser(description="Open a saved HDF5 model in the 3D viewer without recomputing.")
     parser.add_argument("h5_path", nargs="?", help="Path to the HDF5 model file (positional).")
